starting pt for demo

+++ foo.rb
@@ -1,20 +1,35 @@
-require 'rubygems'
+#My first Ruby code
 require 'sinatra'
 require 'fileutils'
 require 'redcarpet'
 
-#set :public_folder, File.dirname(__FILE__)
-
 get '/' do
-    redirect 'index.html'
+  redirect 'index.html'
+end
+
+get '/readme' do
+  redirect 'README.html'
 end
 
 get '/:path.html' do |path|
+
+  mdPath = "./#{path}.md"
+
+  if File.exists?(mdPath) then
+
     content_type 'text/html', :charset => 'utf-8'
 
-    mdPath = "./#{path}.md"
     mdContent = File.open(mdPath, 'r')
     markdown = Redcarpet::Markdown.new(Redcarpet::Render::HTML, extensions = {})
     markdown.render(mdContent.read)
-    
+
+
+  else
+
+    status 404
+    "Not found"
+
+  end
+
 end
+
